fix fn calls to cover redis update to v0.24

+++ crates/controller/src/api/signaling/ws_modules/breakout/storage.rs
@@ -53,11 +53,7 @@ pub async fn set_config(
 ) -> Result<()> {
     if let Some(duration) = config.duration {
         redis_conn
-            .set_ex(
-                BreakoutRoomConfig { room },
-                config,
-                duration.as_secs() as usize,
-            )
+            .set_ex(BreakoutRoomConfig { room }, config, duration.as_secs())
             .await
             .context("Failed to set breakout-room config")
     } else {
